+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 ## Next
 
 * Remove update info from help screen since `update-notifier` includes this already.
-  * [b8696996af49eb5a5c8d74df86a5bc9169fa50b4](https://github.com/mhkeller/kestrel-cli/commit/b8696996af49eb5a5c8d74df86a5bc9169fa50b4)
+  * [3c1cbbf2e2c6ed91568ac585b202fc8a90e98948](https://github.com/mhkeller/kestrel-cli/commit/3c1cbbf2e2c6ed91568ac585b202fc8a90e98948)
 
 ## 1.2.8